separate out deviceplugin API into own repository
