+++ pkg/kubelet/cm/devicemanager/device_plugin_stub.go
@@ -27,7 +27,7 @@ import (
 
 	"google.golang.org/grpc"
 
-	pluginapi "k8s.io/kubernetes/pkg/kubelet/apis/deviceplugin/v1beta1"
+	pluginapi "k8s.io/deviceplugin-api/pkg/apis/deviceplugin/v1beta1"
 	watcherapi "k8s.io/kubernetes/pkg/kubelet/apis/pluginregistration/v1"
 )
