+++ pkg/kubelet/cm/devicemanager/endpoint_test.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	pluginapi "k8s.io/kubernetes/pkg/kubelet/apis/deviceplugin/v1beta1"
+	pluginapi "k8s.io/kubernetes/staging/src/k8s.io/deviceplugin-api/pkg/apis/deviceplugin/v1beta1"
 )
 
 var (
