+++ staging/src/k8s.io/deviceplugin-api/pkg/apis/deviceplugin/v1alpha/constants.go
