+++ staging/src/k8s.io/deviceplugin-api/pkg/apis/deviceplugin/v1beta1/BUILD
